fixes #592 - Using portPropertyFile generates a file with no properties
The parameter portPropertyFile is not honored. Basically when I use it,
I end up with an empty properties file.

The reason for this is that in the current code the PortMapping code is
created in two areas using the runService.getPortMapping method.

1. When we start the docker images which makes sense (because we do all the
logic of finding out the new ports and ip addresses).

2. Just before exposing the container properties. Here it does not make sense
to re-create it, because all the dynamic properties get lost, and as a result,
no properties are written to the properties file (as far as I can see only
the dynamic properties are written to the property file).

The proposed fix is simple: only create once the PortMapping objects and add
them to the PropertyWriteHelper once the images has been started (and thus
we assume that the dynamic assigned ports work properly).